:ambulance: move JetShopeeApp.kt to package screen
