+++ app/src/main/java/com/rivaldy/id/compose/MainActivity.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
+import com.rivaldy.id.compose.ui.screen.JetShopee
 import com.rivaldy.id.compose.ui.theme.JetShopeeTheme
 import com.rivaldy.id.compose.ui.theme.StatusBarColor
 import dagger.hilt.android.AndroidEntryPoint
